requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
scikit-learn==0.24.2
numpy==1.21.2
pandas==1.3.3
schedule==1.1.0
pytest==7.4.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0 